Added bottom bar with new diamond form
Added the diamond form on the bottom of the screen. Currently it does have a price input.

I'm thinking we'll need to add one column to the diamond data that is something like "isCustomDiamond" which will be how we get it to show as a different color on our graphs.